serverSide/requirements.txt
pymysql==0.9.3
mysql-to-sqlite3
torf-cli